net: move CNetAddr/CService/CSubNet out of netbase
